requirements.txt
# Core dependencies
requests>=2.28.0
python-dotenv>=1.0.0
feedparser>=6.0.0

# YouTube API
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.100.0

# Video editing
moviepy>=2.0.0

# Image generation (Stable Diffusion)
torch>=2.0.0
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.0
